requirements.txt
djangorestframework
mongoengine==0.18.2
python-dateutil
Pyjwt
Django==2.2.4
django-cors-headers
nose
pymongo==3.9.0
six==1.10.0
flake8
flake8-import-order
Sphinx==1.5.5
sphinx-rtd-theme==0.2.4
openpyxl==2.5.6
django-crontab==0.7.1
Pillow==5.2.0
face_recognition==1.2.3
dlib==19.19.0
numpy==1.17.2
Click==7.0
face-recognition-models==0.3.0